requirements.txt
uvicorn
fastapi
jinja2
psycopg2-binary
sqlalchemy
python-multipart